[supervisor] add ide ready duration metrics
