+++ install/installer/pkg/components/ide-metrics/configmap.go
@@ -235,6 +235,17 @@ func configmap(ctx *common.RenderContext) ([]runtime.Object, error) {
 			},
 			Buckets: []float64{.005, .025, .05, .1, .5, 1, 2.5, 5, 30, 60, 120, 240, 600},
 		},
+		{
+			Name:    "supervisor_ide_ready_duration_total",
+			Help:    "the IDE startup time",
+			Buckets: []float64{0.1, 0.5, 1, 5, 10, 30, 60},
+			Labels: []config.LabelAllowList{
+				{
+					Name:        "kind",
+					AllowValues: []string{"web", "desktop"},
+				},
+			},
+		},
 	}
 
 	errorReporting := config.ErrorReportingConfiguration{
